Move and rename DFU BLE transport files.
